Move randomString to main
